feat: typescript migration: information left panel
