+++ frontend/src/modules/Workplace/information/Divider.tsx
@@ -1,6 +1,5 @@
 import { styled } from "@mui/material/styles";
 
 export const Divider = styled("div")(() => ({
-    borderTop: "solid 1px #393939",
-  }));
-  
+  borderTop: "solid 1px #393939",
+}));
