+++ frontend/src/modules/Workplace/information/DrawerHeader.tsx
@@ -1,0 +1,9 @@
+import { styled } from "@mui/material/styles";
+
+export const DrawerHeader = styled("div")(({ theme }) => ({
+  display: "flex",
+  alignItems: "center",
+  justifyContent: "space-between",
+  padding: theme.spacing(0, 2),
+  ...theme.mixins.toolbar,
+}));
