--- frontend/src/modules/Workplace/information/ModelProgressBar/index.jsx
@@ -1,69 +1,0 @@
-/*
-    Copyright (c) 2022 IBM Corp.
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-import * as React from "react";
-import PropTypes from "prop-types";
-import { useSelector } from "react-redux";
-import LinearProgress from "@mui/material/LinearProgress";
-import Typography from "@mui/material/Typography";
-import Box from "@mui/material/Box";
-
-function LinearProgressWithLabel(props) {
-  return (
-    <Box sx={{ display: "flex", alignItems: "center", pl: "16px", pr: "16px" }}>
-      <Box sx={{ width: "100%", mr: 1 }}>
-        <LinearProgress variant="determinate" {...props} />
-      </Box>
-      <Box sx={{ minWidth: 35 }}>
-        <Typography
-          variant="body2"
-          color="text.secondary"
-          style={{
-            textAlign: "right",
-            color: "#fff",
-            fontSize: "12px",
-          }}
-        >{`${Math.round(props.value)}%`}</Typography>
-      </Box>
-    </Box>
-  );
-}
-
-LinearProgressWithLabel.propTypes = {
-  /**
-   * The value of the progress indicator for the determinate and buffer variants.
-   * Value between 0 and 100.
-   */
-  value: PropTypes.number.isRequired,
-};
-
-export default function LinearWithValueLabel() {
-  const modelUpdateProgress = useSelector((state) => state.workspace.modelUpdateProgress);
-
-  //   React.useEffect(() => {
-  //     const timer = setInterval(() => {
-  //       setProgress((prevProgress) => (prevProgress >= 100 ? 10 : prevProgress + 10));
-  //     }, 800);
-  //     return () => {
-  //       clearInterval(timer);
-  //     };
-  //   }, []);
-
-  return (
-    <Box sx={{ width: "100%" }}>
-      <LinearProgressWithLabel value={modelUpdateProgress} />
-    </Box>
-  );
-}
