+++ frontend/src/modules/Workplace/information/StatsContainer.tsx
