+++ frontend/src/modules/Workplace/information/TabPanel.tsx
@@ -1,0 +1,22 @@
+import { Box } from "@mui/material";
+
+interface TabPanelProps {
+  children: React.ReactElement;
+  value: number;
+  index: number;
+  className: string;
+}
+
+export const TabPanel = ({ children, value, index, ...other }: TabPanelProps) => {
+  return (
+    <div
+      role="tabpanel"
+      hidden={value !== index}
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      {...other}
+    >
+      {value === index && <Box sx={{ p: 3 }}>{children}</Box>}
+    </div>
+  );
+};
